requirements.txt
scrapy==2.5.1
pandas==1.3.3
psycopg2-binary==2.9.1
SQLAlchemy==1.4.25